Reports and Manual Invoke working
